[hardware] Fix reshuffle conditions
For the reshuffle decision of vd, ara_req.vl should be used instead of car_vl_q. For example, the vmv.sx instruction only writes back one element, not all elements, so the reshuffle cannot be omitted. The ara_req.vl signal contains this information.
The mask for the reshuffle decision should also ensure that the corresponding register is used. For instance, in the vmv.xs instruction, if  rd and vs2 are the same, reshuffling vs2 is still necessary.